+++ src/mocks.ts
@@ -13,6 +13,7 @@ export * from "./models/overtimeCarryover.mocks.js";
 export * from "./models/overtimeReduction.mocks.js";
 export * from "./models/project.mocks.js";
 // export * from "./models/service.mocks.js";
+export * from "./models/surchargeModel.mocks.js";
 export * from "./models/targethours.mocks.js";
 export * from "./models/team.mocks.js";
 export * from "./models/user.mocks.js";